Allow email login & adjust top nav-bar to fix issue #3
Modified login/index.jsx page to accept email as a form of logging in.
Also changed it so that only the username is stored in the cookie. To
avoid breaking nav-bar CSS due to a long username/password, altered
it so that it only ever displays the username (regardless of if
username or email was used to log in), and adjusted styling so that
it can fit a username of max width (16 chars). Also fixed minor typos.