Improve styling for short pages and fix issue #14
Changed the content of the body of the pages to be flex boxes so that they
automatically take up the entire viewport if the page is short (such as
the email verified pages), and if the pages is long, the body content does
not spill over into the footer. This fixes issue #14.

Also began working on feature to resend email verification when requested
by customer on the backend.