Resending verification emails, revamped password reset pages
Resend email verification: Added remote method /customers/{id}/resendVerifyEmail
that allows customers to request a verification email to be resent. Added
corresponding interface on front end "Forgot Password" page to allow users to
request the resending of this email.

Password reset: Revamped the process so that users no longer provide an email
and their new password on the forgot password page. Users must now provide
either their username or email, and then a reset password link is sent to the
email on file. The backend creates a temporary reset password token for the user
to use, and clicking the reset password link brings them to a newly created
page to reset their password to a new value.