Admin page bugfixes, loading icon during registration
Made various fixes to the admin dashboard page to fix issue #10 and
fix issue #11. During testing, discovered that issue #10 actually is not the
real issue, so opened a revised issue #19 that describes the issue more
accurately. The admin dashboard currently cannot view files specific to
a particular user.

Also, added a loading gif that replaces the "Sign Up" button once a prospective
new user tries to sign up. This loading icon kicks in while the front end
is making API requests to the backend to check the validity of the username,
email, and password, as well as creating an account. This fixes issue #17.

Added a message for users when they have no uploaded files or orders placed.
This fixes issue #18.

Added a redirection route for the root page "/". The root page will now
automatically redirect to "/home".

Updated README with these fixes. Also fixes to various typos.