Add /allItems page for listing all products sold
To support the ability to sell both regular products and chip orders, we have
to add a overall "products" page that lists all the products we sell. The
implemented page uses the js-buy-sdk from Shopify to retrieve product data
from our Shopify store and then displays it on the /allItems page.